update CosmosDbStorage and CosmosDbKeyEscape

+++ libraries/botbuilder-azure/src/cosmosDbStorage.ts
@@ -9,7 +9,7 @@
 import { Storage, StoreItems } from 'botbuilder';
 import { ConnectionPolicy, DocumentClient, RequestOptions, UriFactory } from 'documentdb';
 import * as semaphore from 'semaphore';
-import { CosmosDBKeyEscape } from './cosmosDbKeyEscape';
+import { CosmosDbKeyEscape } from './cosmosDbKeyEscape';
 
 const _semaphore: semaphore.Semaphore = semaphore(1);
 
@@ -138,7 +138,7 @@ export class CosmosDbStorage implements Storage {
             value: string;
         }[] = keys.map((key: string, ix: number) => ({
             name: `@id${ix}`,
-            value: CosmosDBKeyEscape.escapeKey(key)
+            value: CosmosDbKeyEscape.escapeKey(key)
         }));
 
         const querySpec: {
@@ -183,8 +183,8 @@ export class CosmosDbStorage implements Storage {
     }
 
     public write(changes: StoreItems): Promise<void> {
-        if (!changes) {
-            throw new Error('Please provide a StoreItems with changes to persist.');
+        if (!changes || Object.keys(changes).length === 0) {
+            return Promise.resolve();
         }
 
         return this.ensureCollectionExists().then(() => {
@@ -195,7 +195,7 @@ export class CosmosDbStorage implements Storage {
                 // The ETag information is updated as an _etag attribute in the document metadata.
                 delete changesCopy.eTag;
                 const documentChange: DocumentStoreItem = {
-                    id: CosmosDBKeyEscape.escapeKey(k),
+                    id: CosmosDbKeyEscape.escapeKey(k),
                     realId: k,
                     document: changesCopy
                 };
@@ -228,11 +228,15 @@ export class CosmosDbStorage implements Storage {
     }
 
     public delete(keys: string[]): Promise<void> {
+        if (!keys || keys.length === 0) {
+            return Promise.resolve();
+        }
+
         return this.ensureCollectionExists().then(() =>
             Promise.all(keys.map((k: string) =>
                 new Promise((resolve: any, reject: any): void =>
                     this.client.deleteDocument(
-                        UriFactory.createDocumentUri(this.settings.databaseId, this.settings.collectionId, CosmosDBKeyEscape.escapeKey(k)),
+                        UriFactory.createDocumentUri(this.settings.databaseId, this.settings.collectionId, CosmosDbKeyEscape.escapeKey(k)),
                         (err: any, data: any): void =>
                             err && err.code !== 404 ? reject(err) : resolve()
                         )
@@ -326,14 +330,4 @@ function getOrCreateCollection(client: DocumentClient,
             });
         });
     });
-}
-
-/**
- * @private
- * Converts the key into a DocumentID that can be used safely with CosmosDB.
- * The following characters are restricted and cannot be used in the Id property: '/', '\', '?', '#'
- * More information at https://docs.microsoft.com/en-us/dotnet/api/microsoft.azure.documents.resource.id?view=azure-dotnet#remarks
- */
-function sanitizeKey(key: string): string {
-    return CosmosDBKeyEscape.escapeKey(key);
 }
